refactor(search): split compose UI into ownership files
